[feat] trainer: per-group LR logging, timing metrics, setup_optimizer_and_scheduler refactor
Training improvements:
- train_starvla.py, train_starvla_cotrain.py, train_starvlm.py: log LR per
  optimizer param group (metrics['learning_rate/{group_name}']) for better
  monitoring in W&B / TensorBoard
- train_starvla.py, train_starvla_cotrain.py: add timing/data and timing/model
  metrics for profiling data-loading vs forward-pass bottlenecks
- trainer_tools.py: refactor setup_optimizer_and_scheduler into shared utility
  (previously duplicated across train scripts)
